+++ app/src/main/java/com/example/pbl2021timerapp/db/time/TimeRoomDatabase.java
@@ -1,4 +1,4 @@
-package com.example.pbl2021timerapp;
+package com.example.pbl2021timerapp.db.time;
 
 import android.content.Context;
