+++ app/src/main/java/com/example/pbl2021timerapp/view/set_time/TimeRecyclerViewAdapter.java
@@ -1,4 +1,4 @@
-package com.example.pbl2021timerapp;
+package com.example.pbl2021timerapp.view.set_time;
 
 import android.view.LayoutInflater;
 import android.view.View;
@@ -6,6 +6,9 @@ import android.view.ViewGroup;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
+
+import com.example.pbl2021timerapp.R;
+import com.example.pbl2021timerapp.db.time.Time;
 
 import java.util.List;
