+++ app/src/main/res/layout/activity_set_time.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SetTimeActivity">
+    tools:context=".view.set_time.SetTimeActivity">
 
     <Button
         android:id="@+id/setTimerButton"
