+++ app/src/main/res/layout/fragment_time_picker.xml
@@ -3,5 +3,5 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TimePickerFragment">
+    tools:context=".view.set_time.TimePickerFragment">
 </FrameLayout>
